fix: use accordion state
